requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
ddgs>=0.1.0
streamlit>=1.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.32.0